v0.10

+++ app/src/main/java/com/artiline/antoon/Database/PrimaryKeyGenerator.java
@@ -1,9 +1,9 @@
 package com.artiline.antoon.Database;
 
 public class PrimaryKeyGenerator {
-    private static long primaryKey = 0;
+    private static int primaryKey = 0;
 
-    public static synchronized long generatePrimaryKey(){
+    public static synchronized int generatePrimaryKey(){
         return primaryKey++;
     }
 }
